+++ collective/auditlog/db.py
@@ -1,13 +1,16 @@
+# coding=utf-8
+from json import loads
 from plone.registry.interfaces import IRegistry
 from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker, scoped_session
+from sqlalchemy.orm import scoped_session
+from sqlalchemy.orm import sessionmaker
 from zope.component import getUtility
 from zope.globalrequest import getRequest
 
 
-def getEngine(conn_string=None, req=None):
+def getEngine(conn_string=None, conn_parameters=None, req=None):
     """
-    cache this on the request object
+    Cache this on the request object
     """
     if req is None:
         req = getRequest()
@@ -17,7 +20,13 @@ def getEngine(conn_string=None, req=None):
         if conn_string is None:
             registry = getUtility(IRegistry)
             conn_string = registry['collective.auditlog.interfaces.IAuditLogSettings.connectionstring']  # noqa
-        engine = create_engine(conn_string)
+        if conn_parameters is None:
+            conn_parameters = registry['collective.auditlog.interfaces.IAuditLogSettings.connectionparameters']  # noqa
+        if not conn_parameters:
+            conn_parameters = {}
+        elif isinstance(conn_parameters, basestring):
+            conn_parameters = loads(conn_parameters)
+        engine = create_engine(conn_string, **conn_parameters)
         if req:
             req.environ['sa.engine'] = engine
     return engine